Throttle on the client side and maintain responsiveness of the servos
Disabling the servos between calls saves power but means we have to
guess how long they'll take to get to where the last command sent them.
By that time we've probably got a new place we want them to be anyway so
there's no point in getting to the old place.